+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Added verbose output. Closes [#37](https://github.com/raphaelstolt/lean-package-validator/issues/37).
+
 ### Fixed
 - Empty glob pattern is catched as invalid. Closes [#38](https://github.com/raphaelstolt/lean-package-validator/issues/38).